requirements.txt
absl-py==0.1.11
astor==0.6.2
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
dataclasses==0.4
decorator==4.2.1
django-extensions
django-dia
djangotoolbox==1.8.0
djongo==1.2.23
docopt==0.6.2
gast==0.2.0
grpcio==1.10.0
html5lib==0.9999999
idna==2.6
Markdown==2.6.11
mongoengine==0.15.0
networkx==2.1
numpy==1.14.1
Pillow==5.0.0
pipreqs==0.4.9
protobuf==3.5.2
pymongo==3.6.1
pytz==2018.3
PyWavelets==0.5.2
requests==2.18.4
scikit-image==0.13.1
scipy==1.0.1
six==1.11.0
sqlparse==0.2.4
tensorboard==1.6.0
termcolor==1.1.0
urllib3==1.22
Werkzeug==0.14.1
wincertstore==0.2
yarg==0.1.9
